requirements.txt
fastapi~=0.88.0
uvicorn
elasticsearch[async]~=8.5.3
aiohttp~=3.8.3
asyncio~=3.4.3
xmltodict~=0.13.0
fastapi-users[sqlalchemy]
aiosqlite
SQLAlchemy~=1.4.46
pydantic~=1.10.4
python-dotenv~=0.21.0